@@ -1,22 +1,22 @@
-module github.com/xen0n/go-workwx
+module github.com/xen0n/go-workwx/v2
 
-go 1.17
+go 1.19
 
 require (
 	github.com/PuerkitoBio/goquery v1.8.1
-	github.com/cenkalti/backoff/v4 v4.1.3
+	github.com/cenkalti/backoff/v4 v4.2.1
 	github.com/russross/blackfriday/v2 v2.1.0
-	github.com/smartystreets/goconvey v1.7.2
-	github.com/urfave/cli/v2 v2.24.4
-	golang.org/x/net v0.11.0
+	github.com/smartystreets/goconvey v1.8.1
+	github.com/urfave/cli/v2 v2.25.7
+	golang.org/x/net v0.12.0
 )
 
 require (
 	github.com/andybalholm/cascadia v1.3.1 // indirect
 	github.com/cpuguy83/go-md2man/v2 v2.0.2 // indirect
-	github.com/gopherjs/gopherjs v0.0.0-20181017120253-0766667cb4d1 // indirect
+	github.com/gopherjs/gopherjs v1.17.2 // indirect
 	github.com/jtolds/gls v4.20.0+incompatible // indirect
-	github.com/smartystreets/assertions v1.2.0 // indirect
+	github.com/smarty/assertions v1.15.0 // indirect
 	github.com/xrash/smetrics v0.0.0-20201216005158-039620a65673 // indirect
-	golang.org/x/text v0.10.0 // indirect
+	golang.org/x/text v0.11.0 // indirect
 )
