@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"strconv"
 
-	"github.com/xen0n/go-workwx"
+	"github.com/xen0n/go-workwx/v2"
 
 	"github.com/urfave/cli/v2"
 )
